feat: template to many pages from ui
