--- apps/accounts/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['@repo/eslint-config/next.js'],
-};
