+++ apps/accounts/eslint.config.js
@@ -1,0 +1,4 @@
+import { nextJsConfig } from '@repo/eslint-config/next-js';
+ 
+/** @type {import("eslint").Linter.Config} */
+export default nextJsConfig;
