+++ apps/accounts/package.json
@@ -2,6 +2,7 @@
   "name": "accounts",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev -p 3001 --turbo",
     "build": "next build",
