+++ apps/accounts/src/app/(accounts)/cards/page.tsx
@@ -1,0 +1,7 @@
+import type { JSX } from "react";
+
+export default function Page(): JSX.Element {
+  return (
+    <div>Cards</div>
+  );
+}
