+++ apps/accounts/src/app/(accounts)/layout.tsx
@@ -1,0 +1,9 @@
+import { MainLayout } from "@repo/ui/main-layout";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <MainLayout>
+      {children}
+    </MainLayout>
+  );
+}
